filewatch: split spec CRUD from actual filesystem watch
While not particularly exciting, this moves the data flow to be
in the direction we want: rather than the FileWatch "controller"
reading other entities and using it to create FileWatch objects,
it simply _reads_ from current state and sets up (or tears down)
filesystem notifiers as appropriate.

The actual CRUD for FileWatch objects happens in a new store
subscriber that does the manifest inspection. In the future, this
too will likely get broken apart so that each "target" type just
constructs FileWatch objects as needed along with whatever other
types to do its work.
