+++ internal/engine/subscribers.go
@@ -33,6 +33,7 @@ func ProvideSubscribers(
 	sw *k8swatch.ServiceWatcher,
 	plm *runtimelog.PodLogManager,
 	pfc *portforward.Controller,
+	fsms *fswatch.ManifestSubscriber,
 	fwm *fswatch.WatchManager,
 	bc *BuildController,
 	cc *configs.ConfigsController,
@@ -67,6 +68,7 @@ func ProvideSubscribers(
 		sw,
 		plm,
 		pfc,
+		fsms,
 		fwm,
 		bc,
 		cc,
